Tested all of the notifications to ensure that they work on the app side - should still confirm the live redis works - but then subsequently fixed any items that were not working. This includes the albums not being propogated after accepting as well as users sending friend requests.
